FlaskWebService2/src/requirements.txt
Flask==2.1.0
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
plotly~=5.10.0
tqdm~=4.64.1
wheel~=0.37.1
setuptools~=59.8.0
py~=1.11.0
Cython~=0.29.28
Werkzeug~=2.2.2
pip~=22.3
tensorflow
Pillow